BPUB-1069 refactor - extract common code to be reused
