+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/ethereum/erc20/bizz/BizzPaymentSupport.java
@@ -18,49 +18,22 @@
 package com.generalbytes.batm.server.extensions.extra.ethereum.erc20.bizz;
 
 import com.generalbytes.batm.common.currencies.CryptoCurrency;
-import com.generalbytes.batm.server.extensions.extra.common.PollingPaymentSupport;
-import com.generalbytes.batm.server.extensions.extra.ethereum.etherscan.EtherScan;
-import com.generalbytes.batm.server.extensions.payment.IPaymentSupport;
-import com.generalbytes.batm.server.extensions.payment.PaymentRequest;
-import com.generalbytes.batm.server.extensions.payment.ReceivedAmount;
+import com.generalbytes.batm.server.extensions.extra.common.QueryableWalletPaymentSupport;
 
-import java.math.BigDecimal;
+import java.util.concurrent.TimeUnit;
 
-public class BizzPaymentSupport extends PollingPaymentSupport implements IPaymentSupport {
-    protected EtherScan etherScan = new EtherScan();
-
-    public void poll(PaymentRequest request) {
-        try {
-            ReceivedAmount addressBalance = etherScan.getAddressBalance(request.getAddress(), request.getCryptoCurrency());
-
-            if (addressBalance.getTotalAmountReceived().compareTo(BigDecimal.ZERO) > 0) {
-                log.info("Received: {}, Requested: {}, {}", addressBalance.getTotalAmountReceived(), request.getAmount(), request);
-                if (addressBalance.getTotalAmountReceived().compareTo(request.getAmount()) == 0) {
-                    if (request.getState() == PaymentRequest.STATE_NEW) {
-                        log.info("Amounts matches {}", request);
-                        request.setTxValue(addressBalance.getTotalAmountReceived());
-                        setState(request, PaymentRequest.STATE_SEEN_TRANSACTION);
-                    }
-                    if (addressBalance.getConfirmations() > 0) {
-                        if (request.getState() == PaymentRequest.STATE_SEEN_TRANSACTION) {
-                            setState(request, PaymentRequest.STATE_SEEN_IN_BLOCK_CHAIN);
-                        }
-                        log.info("{} confirmations for {}", addressBalance.getConfirmations(), request);
-                        updateNumberOfConfirmations(request, addressBalance.getConfirmations());
-                    }
-                } else if (request.getState() != PaymentRequest.STATE_TRANSACTION_INVALID) {
-                    log.info("Received amount does not match the requested amount");
-                    setState(request, PaymentRequest.STATE_TRANSACTION_INVALID);
-                }
-            }
-
-        } catch (Exception e) {
-            log.error("", e);
-        }
-    }
-
+public class BizzPaymentSupport extends QueryableWalletPaymentSupport {
     @Override
     protected String getCryptoCurrency() {
         return CryptoCurrency.BIZZ.getCode();
     }
+
+    protected long getPollingPeriodMillis() {
+        return TimeUnit.MINUTES.toMillis(4);
+    }
+
+    protected long getPollingInitialDelayMillis() {
+        return TimeUnit.SECONDS.toMillis(60);
+    }
+
 }
