+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/slp/SlpPaymentSupport.java
@@ -17,21 +17,13 @@
  ************************************************************************************/
 package com.generalbytes.batm.server.extensions.extra.slp;
 
-import com.generalbytes.batm.server.extensions.IGeneratesNewDepositCryptoAddress;
-import com.generalbytes.batm.server.extensions.IQueryableWallet;
-import com.generalbytes.batm.server.extensions.extra.common.PollingPaymentSupport;
+import com.generalbytes.batm.server.extensions.extra.common.QueryableWalletPaymentSupport;
 import com.generalbytes.batm.server.extensions.payment.IPaymentRequestSpecification;
 import com.generalbytes.batm.server.extensions.payment.PaymentRequest;
-import com.generalbytes.batm.server.extensions.payment.ReceivedAmount;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.math.BigDecimal;
 import java.util.concurrent.TimeUnit;
 
-public class SlpPaymentSupport extends PollingPaymentSupport {
-    private static final Logger log = LoggerFactory.getLogger(SlpPaymentSupport.class);
-
+public class SlpPaymentSupport extends QueryableWalletPaymentSupport {
     private final String cryptoCurrency;
     private final int decimals;
 
@@ -52,12 +44,6 @@ public class SlpPaymentSupport extends PollingPaymentSupport {
 
     @Override
     public PaymentRequest createPaymentRequest(IPaymentRequestSpecification spec) {
-        if (!(spec.getWallet() instanceof IGeneratesNewDepositCryptoAddress)) {
-            throw new IllegalArgumentException("Wallet '" + spec.getWallet().getClass() + "' does not implement " + IGeneratesNewDepositCryptoAddress.class.getSimpleName());
-        }
-        if (!(spec.getWallet() instanceof IQueryableWallet)) {
-            throw new IllegalArgumentException("Wallet '" + spec.getWallet().getClass() + "' does not implement " + IQueryableWallet.class.getSimpleName());
-        }
         if (spec.getTotal().stripTrailingZeros().scale() > decimals) {
             throw new IllegalArgumentException(cryptoCurrency + " has " + decimals + " decimals");
         }
@@ -67,45 +53,5 @@ public class SlpPaymentSupport extends PollingPaymentSupport {
     @Override
     protected String getCryptoCurrency() {
         return cryptoCurrency;
-    }
-
-    @Override
-    protected void poll(PaymentRequest request) {
-        try {
-            IQueryableWallet wallet = (IQueryableWallet) request.getWallet();
-            ReceivedAmount receivedAmount = wallet.getReceivedAmount(request.getAddress(), request.getCryptoCurrency());
-            BigDecimal totalReceived = receivedAmount.getTotalAmountReceived();
-            int confirmations = receivedAmount.getConfirmations();
-
-            if (totalReceived.compareTo(BigDecimal.ZERO) == 0) {
-                return;
-            }
-
-            if (totalReceived.compareTo(request.getAmount()) != 0) {
-                log.info("Received amount ({}) does not match the requested amount ({}), {}", totalReceived, request.getAmount(), request);
-                // stop future polling
-                setState(request, PaymentRequest.STATE_TRANSACTION_INVALID);
-                return;
-            }
-
-            // correct amount received
-
-            if (request.getState() == PaymentRequest.STATE_NEW) {
-                log.info("Received: {}, amounts matches. {}", totalReceived, request);
-                request.setTxValue(totalReceived);
-                setState(request, PaymentRequest.STATE_SEEN_TRANSACTION);
-            }
-
-            if (confirmations > 0) {
-                if (request.getState() == PaymentRequest.STATE_SEEN_TRANSACTION) {
-                    log.info("Transaction confirmed. {}", request);
-                    setState(request, PaymentRequest.STATE_SEEN_IN_BLOCK_CHAIN);
-                }
-                updateNumberOfConfirmations(request, confirmations);
-            }
-
-        } catch (Exception e) {
-            log.error("", e);
-        }
     }
 }
